PZWidgetsKit // Tear down components from Material and OfficialFeed views.
