+++ Packages/PZWidgetsKit/Sources/PZWidgetsKit/ViewsForWidgets/ViewComponents/MaterialsAndFeeds/OfficialFeedList4WidgetsView.swift
@@ -9,57 +9,6 @@ import PZInGameEventKit
 import SwiftUI
 import WidgetKit
 
-// MARK: - OfficialFeedWidgetView
-
-@available(watchOS, unavailable)
-public struct OfficialFeedWidgetView<RefreshIntent: AppIntent>: View {
-    // MARK: Lifecycle
-
-    public init(
-        entry: OfficialFeedWidgetEntry,
-        showLeadingBorder: Bool = true,
-        refreshIntent: RefreshIntent?
-    ) {
-        self.entry = entry
-        self.games = entry.games
-        self.showLeadingBorder = showLeadingBorder
-        self.refreshIntent = refreshIntent
-    }
-
-    // MARK: Public
-
-    public var body: some View {
-        OfficialFeedList4WidgetsView(
-            events: entry.events,
-            showLeadingBorder: showLeadingBorder,
-            refreshIntent: refreshIntent
-        )
-        .environment(\.colorScheme, .dark)
-        .pzWidgetContainerBackground(viewConfig: viewConfig)
-    }
-
-    // MARK: Internal
-
-    @Environment(\.widgetFamily) var family: WidgetFamily
-
-    // MARK: Private
-
-    private let entry: OfficialFeedWidgetEntry
-    private let games: Set<Pizza.SupportedGame>
-    private let showLeadingBorder: Bool
-    private let refreshIntent: RefreshIntent?
-
-    private var viewConfig: WidgetViewConfig {
-        var result = WidgetViewConfig()
-        result.randomBackground = false
-        result.selectedBackgrounds = [
-            WidgetBackground.randomNamecardBackground4Games(entry.games),
-        ]
-        result.isDarkModeRespected = true
-        return result
-    }
-}
-
 // MARK: - OfficialFeedList4WidgetsView
 
 @available(watchOS, unavailable)
